nios2: avuart base type casting
Add mapbase type casting for avuart and move avuart.h to include/linux.

+++ linux-2.6.x/arch/nios2/kernel/config.c
@@ -3,44 +3,44 @@
 #include <linux/init.h>
 #include <linux/interrupt.h>
 #include <linux/io.h>
-#include <asm/avuart.h>
+#include <linux/avuart.h>
 
 static struct avalon_uart_platform_uart nios2_uart_platform[] = {
 #ifdef na_uart0
 	{
-		.mapbase	= na_uart0,
-		.irq		= na_uart0_irq,
-		.uartclk	= nasys_clock_freq,
-	},
+	 .mapbase = (unsigned long)na_uart0,
+	 .irq = na_uart0_irq,
+	 .uartclk = nasys_clock_freq,
+	 },
 #endif
 #ifdef na_uart1
 	{
-		.mapbase 	= na_uart1,
-		.irq		= na_uart1_irq,
-		.uartclk	= nasys_clock_freq,
-	},
+	 .mapbase = (unsigned long)na_uart1,
+	 .irq = na_uart1_irq,
+	 .uartclk = nasys_clock_freq,
+	 },
 #endif
 #ifdef na_uart2
 	{
-		.mapbase 	= na_uart2,
-		.irq		= na_uart2_irq,
-		.uartclk	= nasys_clock_freq,
-	},
+	 .mapbase = (unsigned long)na_uart2,
+	 .irq = na_uart2_irq,
+	 .uartclk = nasys_clock_freq,
+	 },
 #endif
 #ifdef na_uart3
 	{
-		.mapbase 	= na_uart3,
-		.irq		= na_uart3_irq,
-		.uartclk	= nasys_clock_freq,
-	},
+	 .mapbase = (unsigned long)na_uart3,
+	 .irq = na_uart3_irq,
+	 .uartclk = nasys_clock_freq,
+	 },
 #endif
-	{ },
+	{},
 };
 
 static struct platform_device nios2_uart = {
-	.name			= "avalon_uart",
-	.id			= 0,
-	.dev.platform_data	= nios2_uart_platform,
+	.name = "avalon_uart",
+	.id = 0,
+	.dev.platform_data = nios2_uart_platform,
 };
 
 static struct platform_device *nios2_devices[] __initdata = {
